♻️ refactor(app): move global stylesheet under app directory
