+++ src/app/styles/global.css
