+++ src/main.tsx
@@ -1,5 +1,5 @@
 // oxlint-disable-next-line import/no-unassigned-import -- side-effect global styles
-import "./styles/global.css";
+import "./app/styles/global.css";
 
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
